* fix tests for unprojection jacobian
